# docs/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Medical card service Api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/patient": {
      "get": {
        "tags": [
          "Patient controller"
        ],
        "summary": "Getting a list of patients",
        "description": "Allows you to get a list of all patients, if there are no patients then an empty list is returned",
        "operationId": "getAllPatients",
        "responses": {
          "405": {
            "description": "The http method is not supported for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          },
          "200": {
            "description": "Successful return",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Patient controller"
        ],
        "summary": "Replace an existing patient",
        "description": "Allows you to replace an entity with the declared one if they have equal IDs",
        "operationId": "updatePatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "405": {
            "description": "The http method is not supported for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found 404 #1 \u003Cbr/\u003ENot found patient 404 #2 \u003Cbr/\u003E",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate oms number error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "Successful return",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error when setting values",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patient controller"
        ],
        "summary": "Create a patient",
        "description": "Allows adding new patients only with a unique compulsory medical insurance policy",
        "operationId": "createPatient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientPersonalInfo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "405": {
            "description": "The http method is not supported for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON request 100 #1 \u003Cbr\u003E Invalid parameter request type #2 ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate oms number error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error when setting values",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          },
          "200": {
            "description": "Successful create patient",
            "content": {
              "application/json": {

              }
            }
          }
        }
      }
    },
...
    "/patient/{patient_id}": {
      "delete": {
        "tags": [
          "Patient controller"
        ],
        "summary": "Delete patient",
        "description": "Allows you to delete a patient with the specified ID if it exists",
        "operationId": "deletePatient",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "The http method is not supported for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          },
          "200": {
            "description": "Successful delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not find patient 404 #1 \u003Cbr/\u003EResource not found 404 #2",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          },
          "400": {
            "description": "Occurs when the type of the passed request parameter is missing 400 #1 \u003Cbr/\u003EInvalid parameter request type 400 #2",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PatientProfile": {
        "required": [
          "birth_date",
          "first_name",
          "gender",
          "last_name",
          "oms_number",
          "patient_id"
        ],
        "type": "object",
        "properties": {
          "last_name": {
            "maxLength": 100,
            "minLength": 2,
            "pattern": "^[A-Za-zА-ЯЁа-яё]+(-[A-Za-zА-ЯЁа-яё]+)*$",
            "type": "string"
          },
          "first_name": {
            "maxLength": 43,
            "minLength": 2,
            "pattern": "^[A-Za-zА-ЯЁа-яё]+(-[A-Za-zА-ЯЁа-яё]+)*$",
            "type": "string"
          },
          "middle_name": {
            "maxLength": 110,
            "minLength": 0,
            "pattern": "^[A-Za-zА-ЯЁа-яё]+(-[A-Za-zА-ЯЁа-яё]+)*$",
            "type": "string"
          },
          "gender": {
            "maxLength": 20,
            "minLength": 5,
            "pattern": "^[A-Za-zА-ЯЁа-яё]+(-[A-Za-zА-ЯЁа-яё]+)*$",
            "type": "string"
          },
          "birth_date": {
            "type": "string",
            "format": "date"
          },
          "oms_number": {
            "maxLength": 16,
            "minLength": 16,
            "pattern": "^\\d{16}$",
            "type": "string"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ResponseError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Indicates the specific cause of the error",
            "example": "Missing following statement: patient_id, type is UUID"
          },
          "errors": {
            "type": "array",
            "description": "Describes additional information about what happened, may be null",
            "example": "Required request parameter id for method parameter type is present but converted to null",
            "items": {
              "type": "string",
              "description": "Describes additional information about what happened, may be null",
              "example": "Required request parameter id for method parameter type is present but converted to null"
            }
          }
        },
        "description": "An error model containing a message about the need and possibility of additional errors"
      },
      "PatientErrorDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Indicates the specific cause of the error",
            "example": "Missing following statement: patient_id, type is UUID"
          },
          "errors": {
            "type": "array",
            "description": "Describes additional information about what happened, may be null",
            "example": "Required request parameter id for method parameter type is present but converted to null",
            "items": {
              "type": "string",
              "description": "Describes additional information about what happened, may be null",
              "example": "Required request parameter id for method parameter type is present but converted to null"
            }
          },
          "patient_id": {
            "type": "string",
            "description": "patient id",
            "format": "uuid",
            "example": "5b222500-c0fc-40ce-b24a-d35ae89deed7"
          },
          "oms_police": {
            "type": "string",
            "description": "patient oms police, 16 numbers",
            "example": "1234567898765432"
          }
        }
      },
      "PatientResponseDto": {
        "type": "object",
        "properties": {
          "patient_id": {
            "type": "string",
            "format": "uuid"
          },
          "last_name": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "middle_name": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "birth_date": {
            "type": "string",
            "format": "date"
          },
          "oms_number": {
            "type": "string"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Violation"
            }
          }
        }
      },
      "Violation": {
        "type": "object",
        "properties": {
          "fieldName": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "PatientPersonalInfo": {
        "required": [
          "birth_date",
          "first_name",
          "gender",
          "last_name",
          "oms_number"
        ],
        "type": "object",
        "properties": {
          "last_name": {
            "maxLength": 100,
            "minLength": 2,
            "pattern": "^[A-Za-zА-ЯЁа-яё]+(-[A-Za-zА-ЯЁа-яё]+)*$",
            "type": "string"
          },
          "first_name": {
            "maxLength": 43,
            "minLength": 2,
            "pattern": "^[A-Za-zА-ЯЁа-яё]+(-[A-Za-zА-ЯЁа-яё]+)*$",
            "type": "string"
          },
          "middle_name": {
            "maxLength": 110,
            "minLength": 0,
            "pattern": "^[A-Za-zА-ЯЁа-яё]+(-[A-Za-zА-ЯЁа-яё]+)*$",
            "type": "string"
          },
          "gender": {
            "maxLength": 20,
            "minLength": 5,
            "pattern": "^[A-Za-zА-ЯЁа-яё]+(-[A-Za-zА-ЯЁа-яё]+)*$",
            "type": "string"
          },
          "birth_date": {
            "type": "string",
            "format": "date"
          },
          "oms_number": {
            "maxLength": 16,
            "minLength": 16,
            "pattern": "^\\d{16}$",
            "type": "string"
          }
        }
      },
...
    }
  }
}